+++ client/lib/src/core/body_converter.dart
@@ -1,0 +1,36 @@
+import 'dart:convert';
+
+import 'package:meta/meta.dart';
+import 'package:restio/src/core/exceptions.dart';
+import 'package:restio/src/core/request/header/media_type.dart';
+
+@immutable
+class BodyConverter {
+  const BodyConverter();
+
+  Future<String> encode<T>(
+    T value,
+    MediaType contentType,
+  ) async {
+    final mimeType = contentType.mimeType;
+
+    if (mimeType == 'application/json') {
+      return json.encode(value);
+    } else {
+      throw RestioException('Content type $mimeType not supported');
+    }
+  }
+
+  Future<T> decode<T>(
+    String source,
+    MediaType contentType,
+  ) async {
+    final mimeType = contentType.mimeType;
+
+    if (mimeType == 'application/json') {
+      return json.decode(source);
+    } else {
+      throw RestioException('Content type $mimeType not supported');
+    }
+  }
+}